Moved csv_factory to io/csv.py and updated usages.
